+++ src/main/java/com/vintly/domain/mail/entity/Mail.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 
 @Getter
 public class Mail {
+
     private String address;
     private String title;
     private String message;
 
-    @Builder
     public Mail(String address, String title, String message){
         this.address = address;
         this.title = title;
